Replace isObject with isPlainObject
