+++ src/schema/constants.ts
@@ -1,15 +1,1 @@
 export const DEFINITION_ROOT = "#/definitions/";
-
-/**
- * Schema Types
- */
-
-export enum DataTypes {
-  STRING = "string",
-  NUMBER = "number",
-  ARRAY = "array",
-  BOOLEAN = "boolean",
-  OBJECT = "object",
-  NULL = "null",
-  INTEGER = "integer"
-}
