+++ src/schema/types.ts
@@ -3,6 +3,20 @@ import isPlainObject from "lodash/isPlainObject";
 import has from "lodash/has";
 import { JSONSchema7, JSONSchema7Definition } from "json-schema";
 
+/**
+ * Schema Types
+ */
+
+export enum DataTypes {
+  STRING = "string",
+  NUMBER = "number",
+  ARRAY = "array",
+  BOOLEAN = "boolean",
+  OBJECT = "object",
+  NULL = "null",
+  INTEGER = "integer"
+}
+
 export type SchemaType =
   | "string"
   | "number"
@@ -11,6 +25,20 @@ export type SchemaType =
   | "integer"
   | "array"
   | "null";
+
+export enum SchemaKeywords {
+  REQUIRED = "required",
+  ENUM = "enum",
+  CONST = "const",
+  FORMAT = "format",
+  MAXIMUM_LENGTH = "maxLength",
+  MINIMUM_LENGTH = "minimumLength",
+  PATTERN = "pattern",
+  MAXIMUM = "maximum",
+  MINIMUM = "minimum",
+  MINIMUM_ITEMS = "minItems",
+  MAXIMUM_ITEMS = "maxItems"
+}
 
 /**
  *  Object type guard array items key
